requirements.txt
acme==2.1.0
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.5
boto3==1.35.21
botocore==1.35.21
cached-property==1.5.2
cachetools==5.5.0
celery==5.4.0
certbot==2.1.0
certbot-nginx==2.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-filter==24.3
django-guardian==2.4.0
django-jsonfield==1.4.1
django-measurement==3.2.4
django-minio-storage==0.5.7
django-model-utils==5.0.0
django-money==3.5.3
django-parler==2.3
django-parler-rest==2.2
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-polymorphic==3.1.0
django-seed==0.3.1
django-storages==1.14.4
django-url-filter==0.3.15
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-guardian==0.3.0
djangorestframework-simplejwt==5.3.1
djangorestframework-word-filter==0.2.0
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
enum-compat==0.0.3
Faker==29.0.0
GDAL==3.6.2
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
josepy==1.13.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
measurement==3.2.2
minio==7.2.8
oauthlib==3.2.2
parsedatetime==2.6
phonenumbers==8.13.45
pillow==10.4.0
pluggy==1.5.0
progressbar2==4.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.1
psycopg2==2.9.5
psycopg2-binary==2.9.9
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pycurl==7.45.2
Pygments==2.14.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
PySimpleSOAP==1.16.2
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rich==13.3.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
swapper==1.4.0
sympy==1.13.3
thefuzz==0.22.1
toposort==1.10
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
zope.interface==7.0.3
